Update .gitignore files for Meta, *.gz, *.exe
Jekyll =~ 0.11.0 decided to start deleting files in
the destination that do not exist in the source.
This means we have to update our rsync script to
make the source releases/ directory look like the
one on http://cola.tuxfamily.org/.

That means we now have lots of .tar.gz and .exe files
sitting around.  Ignore them.

Signed-off-by: David Aguilar <davvid@gmail.com>