downloads: update gentoo instructions
From: dev-zero <tm@dev-zero.ch>
To: git-cola <git-cola@googlegroups.com>
Date: Tue, 8 Jun 2010 06:11:51 -0700 (PDT)
Subject: git-cola on Gentoo

Hi everyone

Just a short notice: git-cola is available in Gentoo
(actually since June 2009).

To install it you just have to run 'emerge cola' It would be
great if someone could put that on the Downloads-page.

Thanks in advance and thanks for that really cool app.